♻️refactor: refactor keyboard component structure
